# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * RLEGS AM Monitoring & Bot Telegram API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  nik: string;
  fullname: string;
  divisi: string;
  activityEndDate: string;
  label: string;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  caName?: string | null;
  /** @nullable */
  lopid?: string | null;
  /** @nullable */
  picName?: string | null;
  /** @nullable */
  activityNotes?: string | null;
}