# lib/api-zod/src/generated/types/activityLabelGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * RLEGS AM Monitoring & Bot Telegram API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLabelGroup {
  label: string;
  count: number;
}