# lib/api-zod/src/generated/types/activityOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * RLEGS AM Monitoring & Bot Telegram API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityAmGroup } from './activityAmGroup';
import type { ActivityItem } from './activityItem';

export interface ActivityOverview {
  totalActivity: number;
  byAm: ActivityAmGroup[];
  activities: ActivityItem[];
}