# lib/api-zod/src/generated/types/publicAmProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * RLEGS AM Monitoring & Bot Telegram API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityDetail } from './activityDetail';
import type { FunnelDetail } from './funnelDetail';
import type { PerformanceSummary } from './performanceSummary';

export interface PublicAmProfile {
  nik: string;
  nama: string;
  slug: string;
  divisi: string;
  witel: string;
  performance?: PerformanceSummary;
  funnel?: FunnelDetail;
  activity?: ActivityDetail;
}